tools/gn_bazel/requirements.txt
watchdog==6.0.0
python-daemon==3.1.2
